+++ cms/context_processors.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from django.utils import lru_cache
+from functools import lru_cache
+
 from django.utils.functional import lazy
 
 from cms.utils.conf import get_cms_setting
@@ -12,7 +13,7 @@ def cms_settings(request):
     """
     from menus.menu_pool import MenuRenderer
 
-    @lru_cache.lru_cache(maxsize=None)
+    @lru_cache(maxsize=None)
     def _get_menu_renderer():
         # We use lru_cache to avoid getting the manager
         # every time this function is called.