+++ cms/management/commands/subcommands/base.py
@@ -7,7 +7,7 @@ from collections import OrderedDict
 from django.core.management.base import BaseCommand, CommandParser
 from django.core.management.color import no_style, color_style
 
-from cms.utils.compat import DJANGO_2_0, DJANGO_2_2
+from cms.utils.compat import DJANGO_2_0, DJANGO_2_2, DJANGO_3_0
 
 
 def add_builtin_arguments(parser):
@@ -37,9 +37,12 @@ def add_builtin_arguments(parser):
         help='Raise on CommandError exceptions')
     parser.add_argument('--no-color', action='store_true', dest='no_color', default=False,
         help="Don't colorize the command output.")
-    if DJANGO_2_2:
+    if DJANGO_2_2 or DJANGO_3_0:
         parser.add_argument('--force-color', action='store_true', dest='force_color', default=False,
             help="Colorize the command output.")
+    if DJANGO_3_0:
+        parser.add_argument('--skip-checks', action='store_true', dest='skip_checks', default=False,
+            help="Skip the checks.")
 
 
 class SubcommandsCommand(BaseCommand):