+++ cms/tests/test_page_admin.py
@@ -11,10 +11,11 @@ from django.http import HttpRequest
 from django.test.html import HTMLParseError, Parser
 from django.test.utils import override_settings
 from django.urls import clear_url_caches
-from django.utils import six
 from django.utils.encoding import force_text
 from django.utils.timezone import now as tz_now
 from django.utils.translation import override as force_language
+
+from six import string_types
 
 from cms import constants
 from cms.admin.pageadmin import PageAdmin
@@ -1042,7 +1043,6 @@ class PageTest(PageTestBase):
                 {
                     'code': 'en',
                     'name': 'English',
-                    'fallbacks': ['fr', 'de'],
                     'public': True,
                     'fallbacks': ['fr']
                 },
@@ -1682,7 +1682,7 @@ class PageTest(PageTestBase):
             document.finalize()
             # Removing ROOT element if it's not necessary
             if len(document.children) == 1:
-                if not isinstance(document.children[0], six.string_types):
+                if not isinstance(document.children[0], string_types):
                     document = document.children[0]
             return document