+++ cms/utils/compat/dj.py
@@ -1,3 +1,7 @@
+import six
+
+from functools import WRAPPER_ASSIGNMENTS
+
 from django.apps import apps
 from django.conf import settings
 
@@ -6,7 +10,7 @@ __all__ = ['is_installed', 'installed_apps', 'get_apps', 'get_app_paths']
 
 # import these directly from Django!
 from django.utils.encoding import (  # nopyflakes
-    force_text as force_unicode, python_2_unicode_compatible,
+    force_text as force_unicode,
 )
 
 try:
@@ -42,6 +46,11 @@ def get_app_paths():
 def get_apps():
     return [app.models_module for app in apps.get_app_configs()]
 
+def available_attrs(fn):
+    if six.PY3:
+        return WRAPPER_ASSIGNMENTS
+    else:
+        return tuple(a for a in WRAPPER_ASSIGNMENTS if hasattr(fn, a))
 
 def get_middleware():
     if getattr(settings, 'MIDDLEWARE', None) is None: