+++ testserver.py
@@ -157,7 +157,7 @@ HELPER_SETTINGS = dict(
 
 def _helper_patch(*args, **kwargs):
     from django.core.management import call_command
-    from djangocms_helper import utils
+    from app_helper import utils
 
     call_command('migrate', run_syncdb=True)
     utils.create_user('normal', 'normal@normal.normal', 'normal', is_staff=True, base_cms_permissions=True,
@@ -165,12 +165,12 @@ def _helper_patch(*args, **kwargs):
 
 
 def run():
-    from djangocms_helper import runner
-    from djangocms_helper import utils
+    from app_helper import runner
+    from app_helper import utils
 
     os.environ.setdefault('DATABASE_URL', 'sqlite://localhost/testdb.sqlite')
 
-    # Patch djangocms_helper to create tables
+    # Patch app_helper to create tables
     utils._create_db = _helper_patch
 
     # we use '.runner()', not '.cms()' nor '.run()' because it does not